Move builders to separate directory
