+++ ogl-engine/stories/asteroids.stories.tsx
@@ -2,14 +2,13 @@ import React from "react";
 import type { StoryFn, Meta } from "@storybook/react";
 import { Styles } from "@kit/theming/style";
 import { merge } from "lodash";
-import { Asteroids } from "@ogl-engine/engine/Asteroids";
+import { Asteroids } from "@ogl-engine/builders/Asteroids";
 import { BaseMesh } from "@ogl-engine/engine/BaseMesh";
 import { assetLoader } from "@ogl-engine/AssetLoader";
 import { PbrMaterial } from "@ogl-engine/materials/pbr/pbr";
 import { entityScale } from "@ui/components/TacticalMap/EntityMesh";
 import models from "@assets/models";
 import type { Engine3D } from "@ogl-engine/engine/engine3d";
-import { Vec2 } from "ogl";
 import { getFPoints } from "@core/archetypes/asteroidField";
 import { noop } from "@fxts/core";
 import { Story3d, story3dMeta } from "./Story3d";
