+++ ui/components/TacticalMap/TacticalMap.tsx
@@ -14,15 +14,15 @@ import type { Sim } from "@core/sim";
 import { contextMenuStore } from "@ui/state/contextMenu";
 import { storageHook } from "@core/hooks";
 import { MouseButton } from "@ogl-engine/Orbit";
-import { Asteroids } from "@ogl-engine/engine/Asteroids";
+import { Asteroids } from "@ogl-engine/builders/Asteroids";
 import type { Destroyable } from "@ogl-engine/types";
 import { Engine3D } from "@ogl-engine/engine/engine3d";
 import { gameStore } from "@ui/state/game";
 import { reaction } from "mobx";
-import { Star } from "@ogl-engine/engine/Star";
+import { Star } from "@ogl-engine/builders/Star";
 import { Light } from "@ogl-engine/engine/Light";
 import type { Entity } from "@core/entity";
-import { SelectionBox } from "@ogl-engine/engine/SelectionBox";
+import { SelectionBox } from "@ogl-engine/builders/SelectionBox";
 import sounds from "@assets/ui/sounds";
 import { transport3D } from "@core/systems/transport3d";
 import type { GameSettings } from "@core/settings";
